requirements.txt
psutil==5.8.0
paramiko==2.7.2
Twisted==21.2.0
pip==21.0.1
cryptography==3.4.6
setuptools==54.0.0
asyncio==3.4.3
aiogram==2.12.1
pytz==2021.1
django==3.2
psycopg2-binary
